[prometheus-thanos] Populated previously empty keys (#136)
* Fix for misnaming the service account.

Instead of looking for "thanos-prometheus-thanos-ruler-token-..."
Thanos ruler will now look for "prometheus-thanos-ruler-token-...".

Signed-off-by: Andrew Neudegg <andrew.neudegg@finbourne.com>

* Populated previously empty keys.

Empty keys have been populated with empty structures of their exepcted type,
i.e. an array is [], object {}, to reduce yaml related warnings when
deploying with helm. Helm raises a warning when a nil value is populated
with a value of another type.

Signed-off-by: Andrew Neudegg <andrew.neudegg@finbourne.com>

* Populated previously empty keys.

Empty keys have been populated with empty structures of their exepcted type,
i.e. an array is [], object {}, to reduce yaml related warnings when
deploying with helm. Helm raises a warning when a nil value is populated
with a value of another type.

Signed-off-by: Andrew Neudegg <andrew.neudegg@finbourne.com>