Retaining timestamp from previous json parse.
The proposed fix in pull-request #257 did not work on our setup.
But I did find an option on fluentd to retain time, even though embedded json is parsed.
We had the exact problem described in issue #199, and this seems to fix it.
Also the time_key "" did not do anything, so removed it.

Signed-off-by: Anders Nyvang <nanders@nanders.com>