[prometheus-thanos] Fix volume templating logic in compact-statefulset
The logic was broken when persistentVolume was not enabled.
It relied on a set of inexplicit conditions that would allow
template.spec.volumes to be rendered and be the last section before the
final "{{- else }}" statement at the end of the template file.

Fixed by introducing a check of compact.persistentVolume.enabled value
at template.spec.volumes, which makes rendering independent of the later
sibblings of template.spec.volumes and their order.

Signed-off-by: Rozcietrzewiacz <jankeso@gmail.com>