[prometheus-thanos] Fix syntax
Fixes a bug where the template incorrectly chomps whitespace on the
right and by that swallows a newline character and makes the rendered
object invalid. When opting in to binary index headers, the broken chart
merges the flag `--experimental.enable-index-header` and the next line,
`ports:` into one: `--experimental.enable-index-headerports:`

This fixes it.

Fixes: #321
Signed-off-by: Björn Marschollek <bjorn.marschollek@skyscanner.net>